// src/app/Models/new-request.ts
export class NewRequestModel {
    name: string;
    address: string;
    price: number;
    // tslint:disable-next-line:variable-name
    start_time: string;
    // tslint:disable-next-line:variable-name
    end_time: string;
    // tslint:disable-next-line:variable-name
    number_of_hour: number;
    longitude: number;
    latitude: number;
    // tslint:disable-next-line:variable-name
    medical_id: number;
}